Feat-PlayerDamage
- Fix-Math
- HP & Mana are now floats
- Enemy now follows the player and attacks
- Added attack coroutines with cooldowns
- Fixed HealthBar & ManaBar

+++ Assets/C# Scripts/IFight.cs
@@ -3,11 +3,11 @@ public interface IFight
     bool CanTakeDamage();
     //bool CanDealDamage(); //maybe?
 
-    void TakeDamage(int dmg);
+    void TakeDamage(float dmg);
 
     void Die();
 
-    void HealHP(int amount);
+    void HealHP(float amount);
 
     //To do:
     //Handle drain HP (not necessarily on this interface though)
